updated styling for nav and similar movies
